requirement.txt
Flask==2.3.3
Flask-Login==0.6.2
psycopg2-binary==2.9.7
python-dotenv==1.0.0
Flask-WTF==1.1.1
email-validator==2.2.0
Werkzeug==2.3.7
